軽いリファクタ
